Issue #21: Rename poker-cards for consistency
