+++ src/App.svelte
@@ -11,14 +11,14 @@
     import Icons from './01-atoms/IconsReference.svelte';
     import Buttons from './02-molecules/ButtonsReference.svelte';
     import Inputs from './02-molecules/InputsReference.svelte';
-    import Cards from './02-molecules/CardsReference.svelte';
+    import PokerCards from './02-molecules/PokerCardsReference.svelte';
     import ParticipantsList from './03-organisms/ParticipantsListReference.svelte';
 
     const referenceSheets = {
         '/atoms/icons': Icons,
         '/molecules/buttons': Buttons,
         '/molecules/inputs': Inputs,
-        '/molecules/cards': Cards,
+        '/molecules/poker-cards': PokerCards,
         '/organisms/participants-list': ParticipantsList,
     };
