+++ src/main.css
@@ -6,6 +6,6 @@
 
 @import '02-molecules/buttons.css';
 @import '02-molecules/inputs.css';
-@import '02-molecules/cards.css';
+@import '02-molecules/poker-cards.css';
 
 @import '03-organisms/participants-list.css';
